rename module files
